requirements.txt
# GPREC IntelliBot - Dependencies
# Install with: pip install -r requirements.txt

# ── Backend ───────────────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12

# ── Gemini AI ─────────────────────────────────────────────────────────────────
google-generativeai==0.8.3

# ── Frontend ──────────────────────────────────────────────────────────────────
streamlit==1.39.0
requests==2.32.3

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
aiofiles==24.1.0
